aws kms types aliased to plugins/aws-v1/kms
