+++ go/appencryption/pkg/kms/static_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/godaddy/asherah/go/securememory/memguard"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -14,6 +15,38 @@ import (
 	"github.com/godaddy/asherah/go/appencryption/internal"
 	"github.com/godaddy/asherah/go/appencryption/pkg/crypto/aead"
 )
+
+var (
+	genericErrorMessage = "generic error message"
+
+	secretFactory = new(memguard.SecretFactory)
+)
+
+type MockCrypto struct {
+	mock.Mock
+}
+
+func (c *MockCrypto) Encrypt(data, key []byte) ([]byte, error) {
+	ret := c.Called(data, key)
+
+	var bytes []byte
+	if b := ret.Get(0); b != nil {
+		bytes = b.([]byte)
+	}
+
+	return bytes, ret.Error(1)
+}
+
+func (c *MockCrypto) Decrypt(data []byte, key []byte) ([]byte, error) {
+	ret := c.Called(data, key)
+
+	var bytes []byte
+	if b := ret.Get(0); b != nil {
+		bytes = b.([]byte)
+	}
+
+	return bytes, ret.Error(1)
+}
 
 func TestStaticKMS_Encrypt(t *testing.T) {
 	crypto := aead.NewAES256GCM()
@@ -43,11 +76,10 @@ func TestStaticKMS_EncryptKey_ReturnsErrorOnFail(t *testing.T) {
 	crypto := new(MockCrypto)
 	crypto.On("Encrypt", mock.AnythingOfType("[]uint8"), mock.AnythingOfType("[]uint8")).
 		Return(nil, errors.New(genericErrorMessage))
-	crypto.On("KeyByteSize").Return(staticKMSKeySize)
 
 	m, err := NewStatic("bbsPfQTZsmwEcSRKND87WpoC9umuuuOo", crypto)
 	if assert.NoError(t, err) {
-		key, err := internal.GenerateKey(secretFactory, time.Now().Unix(), crypto.KeyByteSize())
+		key, err := internal.GenerateKey(secretFactory, time.Now().Unix(), staticKMSKeySize)
 		if assert.NoError(t, err) {
 			_, err := internal.WithKeyFunc(key, func(keyBytes []byte) ([]byte, error) {
 				return m.EncryptKey(context.Background(), keyBytes)
@@ -61,11 +93,10 @@ func TestStaticKMS_DecryptKey_ReturnsErrorOnFail(t *testing.T) {
 	crypto := new(MockCrypto)
 	crypto.On("Decrypt", mock.AnythingOfType("[]uint8"), mock.AnythingOfType("[]uint8")).
 		Return(nil, errors.New(genericErrorMessage))
-	crypto.On("KeyByteSize").Return(staticKMSKeySize)
 
 	m, err := NewStatic("bbsPfQTZsmwEcSRKND87WpoC9umuuuOo", crypto)
 	if assert.NoError(t, err) {
-		key, err := internal.GenerateKey(secretFactory, time.Now().Unix(), crypto.KeyByteSize())
+		key, err := internal.GenerateKey(secretFactory, time.Now().Unix(), staticKMSKeySize)
 		if assert.NoError(t, err) {
 			_, err := internal.WithKeyFunc(key, func(keyBytes []byte) ([]byte, error) {
 				return m.DecryptKey(context.Background(), keyBytes)
