+++ go/appencryption/plugins/aws-v1/kms/aws_test.go
